generator refactoring
