+++ lib/generators/logidze/install/pg/functions/logidze_capture_exception.sql
