+++ lib/generators/logidze/install/pg/functions/logidze_compact_history.sql
