+++ lib/generators/logidze/install/pg/functions/logidze_logger.sql
