+++ lib/generators/logidze/install/pg/functions/logidze_logger_after.sql
