+++ lib/generators/logidze/install/pg/functions/logidze_snapshot.sql
