+++ lib/generators/logidze/install/pg/functions/logidze_version.sql
