+++ lib/generators/logidze/install/pg/templates/hstore.rb.erb
