+++ lib/generators/logidze/install/pg/templates/migration.rb.erb
