+++ lib/generators/logidze/install/pg/templates/migration_fx.rb.erb
