+++ lib/generators/logidze/model/pg/templates/migration.rb.erb
