+++ lib/generators/logidze/model/pg/triggers/logidze.sql
