+++ lib/generators/logidze/model/pg/triggers/logidze_after.sql
