+++ lib/logidze/utils/func_def.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Logidze
+  module Utils
+    class FuncDef < Struct.new(:name, :version, :signature); end
+  end
+end
