+++ spec/spec_helper.rb
@@ -4,6 +4,7 @@ ENV["RAILS_ENV"] = "test"
 
 begin
   require "debug" unless ENV["CI"] == "true"
+  require "debug/open_nonstop"
 rescue LoadError # rubocop:disable Lint/HandleExceptions
 end
 
@@ -40,5 +41,35 @@ RSpec.configure do |config|
     ActiveRecord::Base.connection.rollback_transaction
 
     raise "Migrations are pending: #{ex.metadata[:location]}" if ActiveRecord::Base.connection.migration_context.needs_migration?
+  end
+end
+
+class Thor
+  class Argument #:nodoc:
+    VALID_TYPES = [:numeric, :hash, :array, :string]
+
+    attr_reader :name, :description, :enum, :required, :type, :default, :banner
+    alias_method :human_name, :name
+
+    def initialize(name, options = {})
+      class_name = self.class.name.split("::").last
+
+      type = options[:type]
+
+      raise ArgumentError, "#{class_name} name can't be nil."                         if name.nil?
+      raise ArgumentError, "Type :#{type} is not valid for #{class_name.downcase}s."  if type && !valid_type?(type)
+
+      @name        = name.to_s
+      @description = options[:desc]
+      @required    = options.key?(:required) ? options[:required] : true
+      @type        = (type || :string).to_sym
+      @default     = options[:default]
+      @banner      = options[:banner] || default_banner
+      @enum        = options[:enum]
+
+      # debugger if name == :name
+
+      validate! # Trigger specific validations
+    end
   end
 end
